flask-segmentation-service/requirements.txt
opencv-python
matplotlib
segmentation_models_pytorch
albumentations
Flask
